rename widget.py -> widgets.py
