+++ tests/functional/sample_main.py
@@ -1,7 +1,7 @@
 #!/usr/bin/env python3
 import retui
 from retui.logger import log
-from retui.widget import Pane, TextBox
+from retui.widgets import Pane, TextBox
 
 
 def test(handle_sigint=True, demo_time_s=None, title=None):
