+++ tests/functional/sample_main_float.py
@@ -1,6 +1,6 @@
 import retui
 from retui.logger import log
-from retui.widget import Pane, TextBox
+from retui.widgets import Pane, TextBox
 
 
 def test(handle_sigint=True, demo_time_s=None, title=None):
